Runtime: Fix clamping on initialization
Ensures clamping logic for metadata datables and clamped properties is run (only if InitFromMetaDataTable() is eventually called).

Although this won't run if InitStats (and consequently InitFromMetaDataTable) is not used, or if InitStats() / DefaultStartingData is used with a nullptr DataTable, or granting / creation of AttributeSets is done without calling InitFromMetaDataTable (Lyra for instance).

Tweaked the unit tests to ensure both metadata datable and clamped properties are clamped on initialization.

Fixes #68